Wrap DotApp around DotPlot
Separation of concerns so DotApp includes UI for settings,
inspecting files, etc. DotPlot has only the plot itself. The
surrounding application only has to include DotApp and hook
it up with a few callbacks to connect to where we load files
from, and save files to, the platform or the local computer.
Those connections to the rest of the world are the only parts
that should be outside DotApp.